Finished UI and added game mechanics
Re-organised code, pulled up all state to top level (app.js) and passed them down as props for best practice.

Added basic game mechanics. Keep in mind that this game is for my step-daughter, so minor bugs might occur. Known issue is that if you click 3 or more cards fast enough, it breaks the game.

Other than this, it is functional and a great game for kids.